feat(neuron-ui): add the story of connection status component, and set the network name to 14px
