+++ packages/neuron-ui/src/stories/index.tsx
@@ -1,3 +1,7 @@
+import 'styles/index.scss'
+import './styles.scss'
+import 'utils/i18n'
+import 'utils/loadTheme'
 import './Overview.stories'
 import './Addresses.stories'
 import './History.stories'
@@ -9,5 +13,4 @@ import './NetworkSetting.stories'
 import './PasswordRequest.stories'
 import './TransactionFeePanel.stories'
 import './SyncStatus.stories'
-import 'styles/index.scss'
-import './styles.scss'
+import './NetworkStatus.stories'
